Add database connection and main functionality

+++ Rml25Client/DeviceData.cs
@@ -4,7 +4,7 @@ namespace Rml25Client
 {
 	class DeviceData
 	{
-		public float WaterLevel { get; set; }
+		public int WaterLevel { get; set; }
 		public DateTime DateTime { get; set; }
 	}
 }
